+++ yasa/tests/test_sleepstats.py
@@ -1,6 +1,4 @@
-"""
-Test the functions in the yasa/sleepstats.py file.
-"""
+"""Test the functions in the yasa/sleepstats.py file."""
 import unittest
 import numpy as np
 import pandas as pd