+++ docs/changelog.rst
@@ -9,6 +9,7 @@ v0.4.2 (dev)
 **Enhancements**
 
 a. Artefact and Unscored epochs are now excluded from the calculation of the total sleep time (TST) in :py:func:`yasa.sleep_statistics`. Previously, YASA calculated TST as SPT - WASO, thus including Art and Uns. TST is now calculated as the sum of all REM and NREM sleep in SPT.
+b. Minor speed improvements in :py:class:`yasa.SleepStaging`.
 
 v0.4.1 (March 2021)
 -------------------