+++ docs/changelog.rst
@@ -3,7 +3,7 @@
 What's new
 ##########
 
-v0.4.2 (dev)
+v0.5.0 (dev)
 ------------
 
 **Bugfix**
@@ -19,6 +19,7 @@ The coincidence matrix gives, for each pair of channel, the number of samples th
 
 a. Artefact and Unscored epochs are now excluded from the calculation of the total sleep time (TST) in :py:func:`yasa.sleep_statistics`. Previously, YASA calculated TST as SPT - WASO, thus including Art and Uns. TST is now calculated as the sum of all REM and NREM sleep in SPT.
 b. Minor speed improvements in :py:class:`yasa.SleepStaging`.
+c. Updated dependency to pyRiemann>=0.2.7, which solves the version conflict with scikit-learn (see `issue 33 <https://github.com/raphaelvallat/yasa/issues/33>`_).
 
 v0.4.1 (March 2021)
 -------------------